Get affect of single parameter on system
The added function based on the input selects the configurations which differ in terms of a single parameter (based on the input argument) and outputs relevant information such as accuracy using those configurations and configuration ids.

Use case: lets say I want to see affect of sampling rates. I can input possible values of sampling rates and the column of sampling rate in config file and a reference configuration from which other parameters are to be taken. This function will output accuracy of all the configs corresponding to diff sampling rates then.
This function will be very useful during the Journal article, where I will have to show the differences due to single parameter.